Add weightedCost calculation

+++ src/main/java/com/mapiz/mystore/unit/domain/Conversion.java
@@ -1,5 +1,6 @@
 package com.mapiz.mystore.unit.domain;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -9,5 +10,5 @@ public class Conversion {
   private Long id;
   private Unit fromUnit;
   private Unit toUnit;
-  private double conversionFactor;
+  private BigDecimal conversionFactor;
 }
